fix title bug

+++ scraping.php
@@ -8,7 +8,6 @@ foreach ($toc as $i) {
     $url = "http://www.esgweb.net/Html/Yxzcpstj/$i.htm";
     echo("$url\n");
     list($title, $content) = $a = get_title_content_by_url($url);
-    print_r($a);
     if (empty($title)) {
         throw new Exception("empty title", 1);
     }
@@ -18,11 +17,14 @@ foreach ($toc as $i) {
 function get_title_content_by_url($url)
 {
     // pass 1 get page
-    $ch = curl_init($url);
-    curl_setopt($ch, CURLOPT_RETURNTRANSFER, 1);
-    $body = curl_exec($ch);
     $cache_file = __DIR__.'/cache/'.str_replace('/', '--', $url);
-    file_put_contents($cache_file, $body);
+    if (!file_exists($cache_file)) {
+        echo "fetch\n";
+        $ch = curl_init($url);
+        curl_setopt($ch, CURLOPT_RETURNTRANSFER, 1);
+        $body = curl_exec($ch);
+        file_put_contents($cache_file, $body);
+    }
     $body = file_get_contents($cache_file);
     $body = mb_convert_encoding($body, 'utf8', 'gbk');
 
@@ -31,8 +33,16 @@ function get_title_content_by_url($url)
     preg_match('%<td[^<]*class="content">(.+?)</td>%ism', $body, $matches);
     $content = $matches[1];
     // get title
-    preg_match('%<font size="3">(.+?)<br>%s', $content, $matches);
-    $title = trim(trim($matches[1]), '　');
+    $is_match = preg_match('%<font size="3">(.+?)<br>|</span><br>([^<]+?)<br>%s', $content, $matches);
+    if (!$is_match) {
+        throw new Exception("no title match", 1);
+    }
+    if (isset($matches[2])) {
+        $title = trim($matches[2]);
+    } else {
+        $title = trim($matches[1]);
+    }
+    echo "title: $title\n";
     // chop start
     // todo performance
     $content = preg_replace('%.+?<font size="3">.+?<br>%s', '', $content);
